adn-alert-backend/requirements.txt
fastapi
uvicorn
paho-mqtt
python-multipart
psycopg2-binary
psycopg2